Back-end-LMS/requirements.txt
flask==2.3.3
flask-cors==4.0.0
databricks-sql-connector==2.9.3
python-dotenv==1.0.0